refactor MapContainer WIP, prepare indicator overlay data #66
