+++ app/components/EntityListViewOptions/index.js
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import ButtonFactory from 'components/buttons/ButtonFactory';
 
 const Styled = styled.div`
-  position: relative;
+  position: ${({ isOnMap }) => isOnMap ? 'absolute' : 'relative'};
   z-index: 20;
   margin-left: 10px;
   padding: 10px 0;
