+++ app/containers/ActorView/ActorActivitiesMap.js
@@ -33,7 +33,7 @@ import {
 // import appMessages from 'containers/App/messages';
 import qe from 'utils/quasi-equals';
 // import { hasGroupActors } from 'utils/entities';
-import MapContainer from 'containers/MapContainer';
+import MapContainer from 'containers/MapContainer/MapWrapper';
 import MapOption from 'containers/MapContainer/MapInfoOptions/MapOption';
 import MapKey from 'containers/MapContainer/MapInfoOptions/MapKey';
 // import messages from './messages';
