+++ app/containers/MapContainer/index.js
@@ -1,554 +1,210 @@
 /*
  *
- * EntitiesMap
+ * MapContainer
  *
  */
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import L from 'leaflet';
-import 'proj4leaflet';
-import { ResponsiveContext } from 'grommet';
-import { merge } from 'lodash/object';
+import { Box, Text } from 'grommet';
 
-import { MAP_OPTIONS } from 'themes/config';
+import * as topojson from 'topojson-client';
+// import { FormattedMessage } from 'react-intl';
 
-import qe from 'utils/quasi-equals';
+import countriesTopo from 'data/ne_countries_10m_v5.topo.json';
+import countryPointsJSON from 'data/country-points.json';
 
-import Tooltip from './Tooltip';
-import { scaleColorCount, getCircleLayer } from './utils';
-
-const Styled = styled.div`
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  background: transparent;
-  z-index: 10;
-  overflow: hidden;
+// import appMessages from 'containers/App/messages';
+// import { hasGroupActors } from 'utils/entities';
+import MapWrapper from './MapWrapper';
+import MapOption from './MapInfoOptions/MapOption';
+import MapKey from './MapInfoOptions/MapKey';
+import MapInfoOptions from './MapInfoOptions';
+const MapKeyWrapper = styled((p) => <Box margin={{ horizontal: 'medium', top: 'xsmall', bottom: 'small' }} {...p} />)`
+  max-width: 400px;
 `;
-const Map = styled.div`
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  background: transparent;
-  z-index: 10;
+// import messages from './messages';
+
+const Styled = styled((p) => <Box {...p} />)`
+  z-index: 0;
 `;
+const MapTitle = styled((p) => <Box margin={{ horizontal: 'medium', vertical: 'xsmall' }} {...p} />)``;
+const MapOptions = styled((p) => <Box margin={{ horizontal: 'medium', top: 'small' }} {...p} />)``;
 
-// const PROJ[projection] = {
-//   name: 'Custom',
-//   crs: 'ESRI:54030',
-//   proj4def: '+proj=robin +lon_0=11.7 +x_0=-11.7 +y_0=0 +datum=WGS84 +units=m +no_defs +type=crs',
-//   resolutions: [
-//     65536, 32768, 16384, 8192, 4096, 2048, 1024, 512, 256, 128,
-//   ],
-//   origin: [0, 11.7],
-//   bounds: [[90, -180], [-90, 180]], // [[N, W], [S, E]]
-// };
-
-const getBBox = (bounds, xLat = 0.5, xLon = 180) => {
-  const nw = bounds[0];
-  const se = bounds[1];
-  const n = nw[0]; // 90
-  const w = nw[1]; // -180
-  const s = se[0]; // -90
-  const e = se[1]; // 180
-  const coordinates = [];
-  // South: SE >> SW
-  for (let lon = e; lon >= w; lon -= xLon) {
-    coordinates.push([lon, s]);
-  }
-  // SW >> NW
-  for (let lat = s; lat <= n; lat += xLat) {
-    coordinates.push([w, lat]);
-  }
-  // NW >> NE
-  for (let lon = w; lon <= e; lon += xLon) {
-    coordinates.push([lon, n]);
-  }
-  // NE >> SE
-  for (let lat = n; lat >= s; lat -= xLat) {
-    coordinates.push([e, lat]);
-  }
-
-  return ({
-    type: 'FeatureCollection',
-    features: [
-      {
-        type: 'Feature',
-        geometry: {
-          type: 'Polygon',
-          coordinates: [coordinates],
-        },
-      },
-    ],
-  });
-};
-
-const TOOLTIP_INITIAL = { features: [] };
-
+const getMapInnerWrapper = (fullMap) => fullMap
+  ? styled.div``
+  : styled((p) => <Box margin={{ horizontal: 'medium' }} {...p} />)`
+    position: relative;
+    height: 500px;
+`;
 export function MapContainer({
-  countryFeatures,
-  countryData,
-  locationData,
-  indicator,
+  mapKey = {},
+  mapInfo,
+  mapOptions = [],
+  mapData = {},
   onActorClick,
-  maxValue,
-  includeSecondaryMembers,
-  mapSubject,
-  fitBounds,
-  options = {},
-  projection = 'robinson',
-  styleType,
-  mapId = 'll-map',
-  interactive = true,
-  scrollWheelZoom = false,
-  layerConfig = {},
+  reducePoints,
+  reduceCountryAreas,
+  fullMap,
+  // intl,
 }) {
-  const mapOptions = merge({}, options, MAP_OPTIONS);
-  const customMapProjection = mapOptions.PROJ[projection];
-  const size = React.useContext(ResponsiveContext);
-  const leafletOptions = customMapProjection
-    ? {
-      crs: new L.Proj.CRS(
-        customMapProjection.crs,
-        customMapProjection.proj4def,
-        {
-          resolutions: customMapProjection.resolutions,
-          origin: customMapProjection.origin,
-          bounds: customMapProjection.bounds,
-        },
-      ),
-      // center: mapOptions.CENTER,
-      // zoom: size === 'small' ? mapOptions.ZOOM.MIN : mapOptions.ZOOM.INIT,
-      zoomControl: size !== 'small' && interactive,
-      dragging: interactive,
-      doubleClickZoom: interactive,
-      scrollWheelZoom,
-      minZoom: mapOptions.ZOOM.MIN,
-      maxZoom: mapOptions.ZOOM.MAX,
-      maxBounds: [
-        [mapOptions.BOUNDS.N, mapOptions.BOUNDS.W],
-        [mapOptions.BOUNDS.S, mapOptions.BOUNDS.E],
-      ],
-      continuousWorld: customMapProjection.continuousWorld || false,
-      worldCopyJump: false,
-      attributionControl: false,
-    }
-    : {
-      // center: mapOptions.CENTER,
-      // zoom: size === 'small' ? mapOptions.ZOOM.MIN : mapOptions.ZOOM.INIT,
-      zoomControl: interactive,
-      dragging: interactive,
-      doubleClickZoom: interactive,
-      scrollWheelZoom,
-      minZoom: mapOptions.ZOOM.MIN,
-      maxZoom: mapOptions.ZOOM.MAX,
-      continuousWorld: true,
-      worldCopyJump: false,
-      attributionControl: false,
-    };
-  const [tooltip, setTooltip] = useState(TOOLTIP_INITIAL);
-  const [featureOver, setFeatureOver] = useState(null);
-  const ref = useRef(null);
-  const mapRef = useRef(null);
-  const countryLayerGroupRef = useRef(null);
-  const countryOverlayGroupRef = useRef(null);
-  const locationOverlayGroupRef = useRef(null);
-  const countryTooltipGroupRef = useRef(null);
-  const countryOverGroupRef = useRef(null);
-  const mapEvents = {
-    // resize: () => {
-    //   // console.log('resize')
-    //   setTooltip(null);
-    // },
-    click: () => {
-      // console.log('mapClick')
-      setTooltip(TOOLTIP_INITIAL);
-    },
-    // mouseover: (a, b, c) => {
-    //   console.log('mouseOver', a, b, c)
-    // },
-    // mousemove: (a, b, c) => {
-    //   console.log('mousemove', a, b, c)
-    // },
-    // zoomstart: () => {
-    //   // console.log('zoomstart')
-    //   setTooltip(null);
-    // },
-    // movestart: () => {
-    //   // console.log('movestart')
-    //   setTooltip(null);
-    // },
-    // layeradd: () => {
-    //   // console.log('layerAdd', layer)
-    //   // setTooltip(null)
-    // },
-    // layerremove: () => {
-    //   // console.log('layerremove', layer)
-    //   // setTooltip(null)
-    // },
-  };
-  const onFeatureClick = (e) => {
-    const { feature } = e.sourceTarget;
-    if (e && L.DomEvent) L.DomEvent.stopPropagation(e);
-    if (e && e.containerPoint && feature && feature.tooltip) {
-      const activeTT = tooltip.features.reduce(
-        (active, f) => f.id === feature.id || active,
-        false,
-      );
-      const newFeatures = activeTT
-        // remove
-        ? tooltip.features.reduce(
-          (memo, f) => f.id === feature.id ? memo : [...memo, f],
-          [],
-        )
-        // add
-        : [
-          feature,
-          ...tooltip.features,
-        ];
-      setTooltip({
-        anchor: e.containerPoint,
-        direction: { x: 'left', y: 'top' },
-        features: newFeatures,
-      });
-    }
-  };
-  const onFeatureOver = (e, feature) => {
-    if (e && L.DomEvent) L.DomEvent.stopPropagation(e);
-    if (!e || !feature || feature.id) setFeatureOver(null);
-    if (feature && feature.id) setFeatureOver(feature.id);
-  };
-  // useEffect(() => {
-  //   /**
-  //    * Alert if clicked on outside of element
-  //    */
-  //   function handleClickOutside(event) {
-  //     console.log(event.target)
-  //     // Do nothing if clicking ref's element or descendent elements
-  //     if (!ref.current || ref.current.contains(event.target)) {
-  //       return;
-  //     }
-  //     setTooltip();
-  //   }
-  //   // Bind the event listener
-  //   document.addEventListener('mousedown', handleClickOutside);
-  //   return () => {
-  //     // Unbind the event listener on clean up
-  //     document.removeEventListener('mousedown', handleClickOutside);
-  //   };
-  // }, [ref]);
-  useEffect(() => {
-    mapRef.current = L.map(mapId, leafletOptions).on(mapEvents);
-    // create an orange rectangle
-    if (customMapProjection && customMapProjection.addBBox) {
-      L.geoJSON(getBBox(customMapProjection.bounds), mapOptions.BBOX_STYLE).addTo(mapRef.current);
-    }
-    countryLayerGroupRef.current = L.layerGroup();
-    countryLayerGroupRef.current.addTo(mapRef.current);
-    countryOverlayGroupRef.current = L.layerGroup();
-    countryOverlayGroupRef.current.addTo(mapRef.current);
-    locationOverlayGroupRef.current = L.layerGroup();
-    locationOverlayGroupRef.current.addTo(mapRef.current);
-    countryTooltipGroupRef.current = L.layerGroup();
-    countryTooltipGroupRef.current.addTo(mapRef.current);
-    countryOverGroupRef.current = L.layerGroup();
-    countryOverGroupRef.current.addTo(mapRef.current);
-    //
-    // mapRef.current.on('zoomend', () => {
-    //   setZoom(mapRef.current.getZoom());
-    // });
-    // mapRef.current.on('moveend', () => {
-    //   onMapMove(getNWSE(mapRef.current));
-    // });
-    mapRef.current.setView(
-      mapOptions.CENTER,
-      mapOptions.ZOOM.INIT,
+  const {
+    indicator,
+    indicatorPoints,
+    mapId,
+    projection,
+    mapSubject,
+    circleLayerConfig,
+    hasPointOption,
+    hasPointOverlay,
+    fitBounds,
+    typeLabels,
+    includeSecondaryMembers,
+    scrollWheelZoom,
+  } = mapData;
+  const {
+    keyTitle,
+    isIndicator,
+    unit,
+    maxBinValue,
+  } = mapKey;
+
+  const [showAsPoint, setShowAsPoint] = useState(false);
+
+  const countriesJSON = topojson.feature(
+    countriesTopo,
+    Object.values(countriesTopo.objects)[0],
+  );
+
+  let countryData = null;
+  let locationData = null;
+  let maxValue;
+  let minValue;
+
+  const showPointsOnly = hasPointOption && showAsPoint;
+  if (
+    reducePoints
+      && indicatorPoints
+      && (hasPointOverlay || showPointsOnly)
+  ) {
+    locationData = reducePoints && reducePoints(
+      countryPointsJSON.features,
+      showAsPoint,
     );
-    if (mapRef.current.zoomControl) {
-      mapRef.current.zoomControl.setPosition('topleft');
-    }
-  }, []);
+    [maxValue, minValue] = locationData && locationData.reduce(
+      ([max, min], feature) => ([
+        max !== null ? Math.max(max, feature.values[indicatorPoints]) : feature.values[indicatorPoints],
+        min !== null ? Math.min(min, feature.values[indicatorPoints]) : feature.values[indicatorPoints],
+      ]),
+      [null, null],
+    );
+  }
+  if (
+    reduceCountryAreas
+    && indicator
+    && !showPointsOnly
+  ) {
+    countryData = reduceCountryAreas && reduceCountryAreas(countriesJSON.features);
 
-  // add countryFeatures basemap
-  useEffect(() => {
-    if (countryFeatures) {
-      countryLayerGroupRef.current.clearLayers();
-      const jsonLayer = L.geoJSON(
-        countryFeatures,
-        {
-          style: () => ({
-            ...mapOptions.DEFAULT_STYLE,
-          }),
-        },
-      );
-      countryLayerGroupRef.current.addLayer(jsonLayer);
-    }
-  }, [countryFeatures]);
+    [maxValue, minValue] = countryData
+      ? countryData.reduce(
+        ([max, min], feature) => ([
+          max !== null ? Math.max(max, feature.values[indicator]) : feature.values[indicator],
+          min !== null ? Math.min(min, feature.values[indicator]) : feature.values[indicator],
+        ]),
+        [null, null],
+      )
+      : [0, 0];
+  }
 
-  // add countryData
-  useEffect(() => {
-    countryOverlayGroupRef.current.clearLayers();
-    if (countryData && countryData.length > 0) {
-      const scale = mapSubject
-        && scaleColorCount(maxValue, mapOptions.GRADIENT[mapSubject], indicator === 'indicator');
-      // treat 0 as no data when showing counts
-      const noDataThreshold = indicator === 'indicator' ? 0 : 1;
-      const jsonLayer = L.geoJSON(
-        countryData,
-        {
-          style: (f) => {
-            const defaultStyle = styleType && mapOptions.STYLE[styleType]
-              ? {
-                ...mapOptions.DEFAULT_STYLE,
-                ...mapOptions.STYLE[styleType],
-              }
-              : mapOptions.DEFAULT_STYLE;
-            const fstyle = f.isActive
-              ? {
-                ...defaultStyle,
-                ...mapOptions.STYLE.active,
-              }
-              : defaultStyle;
-            if (mapSubject) {
-              if (
-                f.values
-                && typeof f.values[indicator] !== 'undefined'
-                && f.values[indicator] >= noDataThreshold
-              ) {
-                return {
-                  ...fstyle,
-                  fillColor: scale(f.values[indicator]),
-                  ...f.style,
-                };
-              }
-              return {
-                ...fstyle,
-                fillColor: mapOptions.NO_DATA_COLOR,
-                ...f.style,
-              };
-            }
-            return {
-              ...fstyle,
-              ...f.style,
-            };
-          },
-        },
-      ).on({
-        click: (e) => onFeatureClick(e),
-        mouseout: () => onFeatureOver(),
-      });
-      countryOverlayGroupRef.current.addLayer(jsonLayer);
-    }
-  }, [countryData, indicator, tooltip, mapSubject]);
-  // add zoom to countryData
-  useEffect(() => {
-    if (
-      fitBounds
-      && countryData
-      && countryData.length > 0
-      && countryOverlayGroupRef
-      && countryOverlayGroupRef.current
-      && countryOverlayGroupRef.current.getLayers()
-      && countryOverlayGroupRef.current.getLayers().length > 0
-    ) {
-      const jsonLayer = countryOverlayGroupRef.current.getLayers()[0];
-      if (jsonLayer.getBounds) {
-        const boundsZoom = mapRef.current.getBoundsZoom(
-          jsonLayer.getBounds(),
-          false, // inside,
-          [20, 20], // padding in px
-        );
-        const boundsCenter = jsonLayer.getBounds().getCenter();
-        // add zoom level to account for custom proj issue
-        const ZOOM_OFFSET = 0;
-        const MAX_ZOOM = 7;
-        mapRef.current.setView(
-          boundsCenter,
-          Math.min(
-            Math.max(boundsZoom - ZOOM_OFFSET, 0),
-            MAX_ZOOM,
-          ),
-          {
-            animate: false,
-          },
-        );
-      }
-    }
-  }, [countryData]);
-  // add zoom to locationData
-  useEffect(() => {
-    if (
-      fitBounds
-      && locationData
-      && locationData.length > 0
-      && locationOverlayGroupRef
-      && locationOverlayGroupRef.current
-      && locationOverlayGroupRef.current.getLayers()
-      && locationOverlayGroupRef.current.getLayers().length > 0
-    ) {
-      const jsonLayer = locationOverlayGroupRef.current.getLayers()[0];
-      if (jsonLayer.getBounds) {
-        const boundsZoom = mapRef.current.getBoundsZoom(
-          jsonLayer.getBounds(),
-          false, // inside,
-          [20, 20], // padding in px
-        );
-        const boundsCenter = jsonLayer.getBounds().getCenter();
-        // add zoom level to account for custom proj issue
-        const ZOOM_OFFSET = 0;
-        const MAX_ZOOM = 7;
-        mapRef.current.setView(
-          boundsCenter,
-          Math.min(
-            Math.max(
-              boundsZoom - ZOOM_OFFSET,
-              0,
-            ),
-            MAX_ZOOM,
-          ),
-          {
-            animate: false,
-          },
-        );
-      }
-    }
-  }, [locationData]);
+  let allMapOptions = mapOptions;
+  if (hasPointOption) {
+    allMapOptions = [
+      {
+        active: showAsPoint,
+        onClick: () => setShowAsPoint(!showAsPoint),
+        label: 'Show as circles',
+        key: 'circle',
+      },
+      ...mapOptions,
+    ];
+  }
 
-  // add locationData
-  useEffect(() => {
-    locationOverlayGroupRef.current.clearLayers();
-    if (locationData && locationData.length > 0) {
-      const layer = L.featureGroup(null, { pane: 'overlayPane' });
-      const jsonLayer = getCircleLayer({
-        features: locationData,
-        config: layerConfig,
-        markerEvents: {
-          click: (e) => onFeatureClick(e),
-          mouseout: () => onFeatureOver(),
-        },
-      });
-      layer.addLayer(jsonLayer);
-      locationOverlayGroupRef.current.addLayer(layer);
-    }
-  }, [locationData, indicator, tooltip, mapSubject, layerConfig]);
-
-  // highlight tooltip feature
-  useEffect(() => {
-    countryTooltipGroupRef.current.clearLayers();
-    if (countryData && tooltip && tooltip.features && tooltip.features.length > 0) {
-      tooltip.features.forEach(
-        (ttFeature) => {
-          const jsonLayer = L.geoJSON(
-            countryData.filter((f) => qe(f.id, ttFeature.id)),
-            { style: mapOptions.TOOLTIP_STYLE },
-          );
-          countryTooltipGroupRef.current.addLayer(jsonLayer);
-        }
-      );
-    }
-    if (locationData && locationOverlayGroupRef.current && locationOverlayGroupRef.current.getLayers()) {
-      const layerGroup = locationOverlayGroupRef.current.getLayers()[0];
-      const layer = layerGroup && layerGroup.getLayers()[0];
-      if (layer) {
-        if (tooltip && tooltip.features && tooltip.features.length > 0) {
-          const tooltipFeatureIds = tooltip.features.map((f) => f.id);
-          if (layer && layer.getLayers()) {
-            layer.getLayers().filter(
-              (f) => tooltipFeatureIds.indexOf(f.feature.id) > -1
-            ).forEach(
-              (f) => {
-                f.bringToFront();
-                f.setStyle({ weight: 1.5 });
-              }
-            );
-          }
-        } else {
-          layer.getLayers().forEach(
-            (feature) => feature.setStyle({ weight: 0.5 })
-          );
-        }
-      }
-    }
-  }, [tooltip, mapSubject, includeSecondaryMembers]);
-
-  useEffect(() => {
-    countryOverGroupRef.current.clearLayers();
-    if (featureOver && countryData) {
-      const jsonLayer = L.geoJSON(
-        countryData.filter((f) => qe(f.id, featureOver)),
-        {
-          style: mapOptions.OVER_STYLE,
-        },
-      );
-      countryOverGroupRef.current.addLayer(jsonLayer);
-    }
-  }, [featureOver]);
-  // update tooltip
-  useEffect(() => {
-    if (tooltip && countryData) {
-      if (tooltip.features && tooltip.features.length > 0) {
-        setTooltip({
-          features: tooltip.features.map(
-            (f) => countryData.find((c) => qe(c.id, f.id))
-          ).filter(
-            (f) => !!f
-          ),
-        });
-      } else {
-        setTooltip(TOOLTIP_INITIAL);
-      }
-    } else {
-      setTooltip(TOOLTIP_INITIAL);
-    }
-  }, [mapSubject, countryData]);
-
+  const MapInnerWrapper = getMapInnerWrapper(fullMap);
   return (
     <Styled>
-      <Map id={mapId} ref={ref} styleType={styleType} />
-      {tooltip && tooltip.features && tooltip.features.length > 0 && (
-        <Tooltip
-          isLocationData={!countryData && !!locationData}
-          mapRef={ref}
-          position={null}
-          direction={tooltip.direction}
-          features={tooltip.features && tooltip.features.map((f) => f && f.tooltip)}
-          onFeatureClick={onActorClick ? (id) => onActorClick(id) : null}
-          onClose={(id) => setTooltip({
-            ...tooltip,
-            features: tooltip.features.reduce(
-              (memo, f) => f.id === id ? memo : [...memo, f],
-              [],
-            ),
-          })}
+      <MapInnerWrapper>
+        <MapWrapper
+          scrollWheelZoom={scrollWheelZoom}
+          typeLabels={typeLabels}
+          includeSecondaryMembers={includeSecondaryMembers}
+          countryData={countryData}
+          locationData={locationData}
+          countryFeatures={countriesJSON.features}
+          indicator={indicator}
+          onActorClick={(id) => onActorClick(id)}
+          maxValue={maxValue}
+          mapSubject={mapSubject}
+          fitBounds={fitBounds}
+          projection={projection}
+          mapId={mapId}
+          circleLayerConfig={circleLayerConfig}
         />
+      </MapInnerWrapper>
+      {mapInfo && Object.keys(mapInfo).length > 0 && (
+        <MapInfoOptions
+          config={{
+            ...mapInfo,
+            maxValue,
+          }}
+          mapSubject={mapSubject}
+        />
+      )}
+      {mapKey && Object.keys(mapKey).length > 0 && (
+        <>
+          <MapTitle>
+            <Text weight={600}>{keyTitle}</Text>
+          </MapTitle>
+          <MapKeyWrapper>
+            <MapKey
+              mapSubject={mapSubject}
+              maxValue={maxValue}
+              minValue={minValue}
+              maxBinValue={maxBinValue}
+              isIndicator={isIndicator}
+              type={hasPointOption && showAsPoint ? 'circles' : 'gradient'}
+              unit={unit}
+              circleLayerConfig={circleLayerConfig}
+            />
+          </MapKeyWrapper>
+        </>
+      )}
+      {allMapOptions && allMapOptions.length > 0 && (
+        <MapOptions>
+          {allMapOptions.map(
+            (option, id) => (
+              <MapOption
+                key={id}
+                option={option}
+                type={option.type}
+              />
+            )
+          )}
+        </MapOptions>
       )}
     </Styled>
   );
 }
 
 MapContainer.propTypes = {
-  countryFeatures: PropTypes.array, // country basemap
-  countryData: PropTypes.array, // country data overlay
-  locationData: PropTypes.array, // location data overlay
-  indicator: PropTypes.string,
   onActorClick: PropTypes.func,
-  maxValue: PropTypes.number,
-  includeSecondaryMembers: PropTypes.bool,
-  fitBounds: PropTypes.bool,
-  interactive: PropTypes.bool,
-  scrollWheelZoom: PropTypes.bool,
-  mapSubject: PropTypes.string,
-  projection: PropTypes.string,
-  styleType: PropTypes.string,
-  mapId: PropTypes.string,
-  options: PropTypes.object,
-  layerConfig: PropTypes.object,
-  // onSetMapSubject: PropTypes.func,
+  reducePoints: PropTypes.func,
+  reduceCountryAreas: PropTypes.func,
+  mapData: PropTypes.object,
+  mapKey: PropTypes.object,
+  mapInfo: PropTypes.object,
+  mapOptions: PropTypes.array,
+  fullMap: PropTypes.bool,
 };
 
 export default MapContainer;
